Moved some classes into their own files
