+++ AalTest/AalTest.vcxproj
@@ -102,9 +102,14 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="source\AalTest\Exceptions.h" />
+    <ClInclude Include="source\AalTest\ValueMismatchTestException.h" />
     <ClInclude Include="source\AalTest\Asserts.h" />
     <ClInclude Include="source\AalTest.h" />
+    <ClInclude Include="source\AalTest\FailedTestException.h" />
+    <ClInclude Include="source\AalTest\ParameterizedTest.h" />
+    <ClInclude Include="source\AalTest\SimpleTest.h" />
+    <ClInclude Include="source\AalTest\SkipTestException.h" />
+    <ClInclude Include="source\AalTest\TestBase.h" />
     <ClInclude Include="source\AalTest\TestResult.h" />
     <ClInclude Include="source\AalTest\TestRunner.h" />
     <ClInclude Include="source\AalTest\TestRunnerOutputBase.h" />
@@ -115,8 +120,10 @@
     <ClInclude Include="source\AalTest\TestSuiteResult.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="source\AalTest\Exceptions.cpp" />
+    <ClCompile Include="source\AalTest\ValueMismatchTestException.cpp" />
     <ClCompile Include="source\AalTest\Asserts.cpp" />
+    <ClCompile Include="source\AalTest\FailedTestException.cpp" />
+    <ClCompile Include="source\AalTest\SkipTestException.cpp" />
     <ClCompile Include="source\AalTest\TestRunner.cpp" />
     <ClCompile Include="source\AalTest\TestSuite.cpp" />
     <ClCompile Include="source\AalTest\TestRunnerWindowsConsoleOutput.cpp" />
