+++ AalTest/source/AalTest/FailedTestException.cpp
@@ -1,0 +1,6 @@
+#include "FailedTestException.h"
+
+FailedTestException::FailedTestException(const std::source_location& sourceLocation)
+    : location{ sourceLocation }
+{
+}
