+++ AalTest/source/AalTest/SkipTestException.cpp
@@ -1,0 +1,6 @@
+#include "SkipTestException.h"
+
+SkipTestException::SkipTestException(const std::source_location& sourceLocation)
+    : location{ sourceLocation }
+{
+}
