+++ AalTest/source/AalTest/TestRunnerNullOutput.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <AalTest/Asserts.h>
 #include <AalTest/TestRunnerOutputBase.h>
 
 class TestRunnerNullOutput : public TestRunnerOutputBase
@@ -13,7 +14,7 @@ public:
     void updateTestResult(const QPoint& position, TestResult result) override {};
     void writeTestPassedMessage() override {};
     void writeTestSkippedMessage(SkipTestException& e) override {};
-    void writeTestFailedMessage(FailTestException& e) override {};
+    void writeTestFailedMessage(FailedTestException& e) override {};
     void writeTestValueMismatchMessage(ValueMismatchTestException& e) override {};
     void writeTestRunnerResult(const TestSuiteResult& result) override {};
     void writeTestRunnerTotalResult(const QList<TestSuiteResult>& results) override {};
