+++ AalTest/source/AalTest/TestRunnerWindowsConsoleOutput.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <AalTest/TestRunnerOutputBase.h>
-#include <QString>
 
 class TestRunnerWindowsConsoleOutput : public TestRunnerOutputBase
 {
@@ -14,7 +13,7 @@ public:
     void updateTestResult(const QPoint& position, TestResult result) override;
     void writeTestPassedMessage() override;
     void writeTestSkippedMessage(SkipTestException& e) override;
-    void writeTestFailedMessage(FailTestException& e) override;
+    void writeTestFailedMessage(FailedTestException& e) override;
     void writeTestValueMismatchMessage(ValueMismatchTestException& e) override;
     void writeTestRunnerResult(const TestSuiteResult& result) override;
     void writeTestRunnerTotalResult(const QList<TestSuiteResult>& results) override;
