+++ AalTest/source/AalTest/ValueMismatchTestException.h
@@ -4,22 +4,6 @@
 #include <QString>
 #include <source_location>
 
-class FailTestException : public std::exception
-{
-public:
-    FailTestException(const std::source_location& sourceLocation);
-
-    std::source_location location;
-};
-
-class SkipTestException : public std::exception
-{
-public:
-    SkipTestException(const std::source_location& sourceLocation);
-
-    std::source_location location;
-};
-
 class ValueMismatchTestException : public std::exception
 {
 public:
